snotebook/requirements.txt
asgiref==3.4.0
Django==3.2.4
django-autoslug==1.9.8
django-widget-tweaks==1.4.8
psycopg2==2.9.1
python-dotenv==0.17.1
pytils==0.3
pytz==2021.1
sqlparse==0.4.1
